Add no_extras to bap.utils.run.run_bap_with
This also updates the BAP View output to show the execution string with
all the details including the extras (if any). This was not needed
earlier since the extras were added always, but now what we can (and
will) disable extras once in a while, it becomes imperative to show the
complete information at that point.