it looks like 3.5 includes previous one
Basically tox here competes with travis. Tox is capable of running
tests in multiple environments, and the same feature is provided by
Travis. I would like to use Tox here, to have the same setup as I have
on my local machine. It is possible to use tox-travis, but it will
still have a setup that would be different from the local one.