fixes and enhances function recover plugin
The function recovery plugin was broken for some amount of time, as we
passed symbol file to bap, effectively disabling BAP own function
finding capabilities, so bap always returned the same amount of function
starts.

This commit also enhances the recovery plugin, by lowering the recovery
threshold. Although it will lead to more false positives it is
acceptable since IDA will ignore function starts that occur in a body of
an discovered function, that was already discovered and
disassembled. Since most of the functions have their bodies after
starts, we added sorting of the addresses, so that function starts are
added to the system in the ascending order.

This commit also adds two new attributes to the BapIda class. The first
one is called [args] and is a list of default arguments, shared by all
instances. It is useful to adapt the behavior of the plugins to your
needs, and to perform fast prototyping and debugging.

The second attribute is called [plugins] and it contains a list of
available BAP plugins. This is useful, when an IDA plugin needs to adapt
its behavior based on the presence of particular BAP plugins. For
example, the function start identification plugin, is allowed to tackle
with the byteweight parameters only if it is installed.