updates to BAP 1.4 release
In this update we are relying on the new Primus Taint Analysis
Framework to provide us data-flow information via the
`primus-propagate-taint` compatibility layer. We allow a user to
select the taint propagation engine, as well as its parameters.

This commit also moves the bap_taint module to the utilities package,
as otherwise it is loaded multiple times by each plugin separatly that
may lead to unexpected results (don't ask me what are they)

There is also a small fix that prevents the racing condition between
askXXX dialogs and IDA Python breakability facility.

The merge of this PR is blocked until
https://github.com/BinaryAnalysisPlatform/bap/pull/784 is merged